2021-2022/first-term/software-engineering/mysite/requirements.txt
djangorestframework~=3.12.4
requests~=2.26.0
django-rest-swagger
Django~=3.2.9